reservation-service/requirements.txt
flask
python-dotenv
pytest
psycopg2-binary
requests